requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
asjson==3.0.0
beautifulsoup4==4.10.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.11
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.11.2
django-import-export==2.7.1
django-mathfilters==1.0.0
django-select2==7.10.0
django-xhtml2pdf==0.0.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==2.10
jsondict==1.2
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.1
paypal==1.2.5
Pillow==8.3.2
Pmw==2.0.1
PyPDF2==1.26.0
pyRXP==2.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
reportlab==3.5.67
requests==2.25.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
